+++ src/modules/fundingManager/oracle/interfaces/IOraclePrice_v1.sol
@@ -35,10 +35,10 @@ interface IOraclePrice_v1 {
     /// @notice Gets current price for token issuance (buying tokens)
     /// @return price_ Current price in 18 decimals (collateral tokens per 1 issuance token)
     /// @dev    Example: If price is 2 USDC/ISS, returns 2e18 (2 USDC needed for 1 ISS)
-    function getPriceForIssuance() external view returns (uint256 price_);
+    function getPriceForIssuance() external view returns (uint price_);
 
     /// @notice Gets current price for token redemption (selling tokens)
     /// @return price_ Current price in 18 decimals (collateral tokens per 1 issuance token)
     /// @dev    Example: If price is 1.9 USDC/ISS, returns 1.9e18 (1.9 USDC received for 1 ISS)
-    function getPriceForRedemption() external view returns (uint256 price_);
+    function getPriceForRedemption() external view returns (uint price_);
 }
